refactor reorganize export to resolve ESLint warning
